refactor(connectors): move the data-connector contract into a data-connector-api crate
The component configuration a connector is built for and the errors it
reports lived in the `runtime` crate, so every one of the ~30 connector
crates had to depend on the orchestrator just to name `ConnectorComponent`
or construct a `DataConnectorError`. Neither type needs anything from the
runtime now that the component carries only its spec.

Move `ConnectorComponent`, `DataConnectorError` and the result aliases into
a new `data-connector-api` crate that sits below `runtime` and depends only
on `runtime-component`, `snafu` and `globset`. `runtime::dataconnector`
re-exports them, so no connector changes here; they can drop their `runtime`
dependency once the `DataConnector` traits follow.

`runtime-component` moves from the `runtime` tier to `shared-utility` in
`layers.toml` — it was already eligible (every dependency it has is
shared-utility or lower), and the api crate needs it below `runtime` to name
`DatasetSpec`/`CatalogSpec`.

`From<&Arc<Dataset>> for ConnectorComponent` is gone: with the target type
now foreign to `runtime`, `Arc` is not a fundamental type, so the impl
violates the orphan rule. The two call sites pass `ds.as_ref()` and use the
`&Dataset` impl.
