+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
     "crates/datafusion-ddl",
     "crates/datafusion-dml",
     "crates/datafusion-optimizer-rules",
+    "crates/data-connector-api",
     "crates/data_components",
     "crates/dataformat-json",
     "crates/document_parse",
