+++ crates/data-connector-api/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+description = "The data-connector contract shared by the Spice runtime and every connector: the component configuration a connector is built for, and the errors it reports."
+edition.workspace = true
+exclude.workspace = true
+homepage.workspace = true
+license-file.workspace = true
+name = "data-connector-api"
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+globset = { workspace = true }
+runtime-component = { path = "../runtime-component" }
+snafu.workspace = true
+
+[lints]
+workspace = true
